[fix] : modify setRole method
