+++ src/main/java/kr/where/backend/admin/AdminService.java
@@ -4,7 +4,6 @@ import java.util.List;
 import kr.where.backend.admin.dto.RequestRoleStatusDTO;
 import kr.where.backend.admin.dto.ResponseRoleStatusDTO;
 import kr.where.backend.admin.dto.ResponseRoleStatusListDTO;
-import kr.where.backend.api.exception.RequestException;
 import kr.where.backend.auth.authUser.AuthUser;
 import kr.where.backend.member.Member;
 import kr.where.backend.member.MemberRepository;
@@ -16,8 +15,6 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class AdminService {
     private final MemberRepository memberRepository;
-    private static final String ADMIN_ROLE = "ADMIN";
-    private static final String USER_ROLE = "USER";
 
     public ResponseRoleStatusDTO getRoleStatus(final AuthUser authUser) {
         final Member member = memberRepository.findByIntraId(authUser.getIntraId())
@@ -27,19 +24,11 @@ public class AdminService {
     }
 
     public ResponseRoleStatusDTO changeAdminStatus(final RequestRoleStatusDTO requestRoleStatusDTO) {
-        validateRole(requestRoleStatusDTO.getRole());
-
         Member targerMember = memberRepository.findByIntraName(requestRoleStatusDTO.getIntraName())
                 .orElseThrow(MemberException.NoMemberException::new);
         if (!targerMember.getRole().equals(requestRoleStatusDTO.getRole()))
-            targerMember.setRole(requestRoleStatusDTO.getRole());
+            targerMember.updateRole(requestRoleStatusDTO.getRole());
         return new ResponseRoleStatusDTO(targerMember.getIntraName(), targerMember.getRole());
-    }
-
-    private void validateRole(final String role) {
-        if (!role.equals(ADMIN_ROLE) && !role.equals(USER_ROLE)) {
-            throw new RequestException.BadRequestException();
-        }
     }
 
     public ResponseRoleStatusListDTO getAllAdmin() {
